Move HealthModel service URL/api-version tests into EntityGetHealthCommandTests
