+++ tools/Azure.Mcp.Tools.Monitor/tests/Azure.Mcp.Tools.Monitor.Tests/HealthModels/Entity/EntityGetHealthCommandTests.cs
@@ -2,11 +2,16 @@
 // Licensed under the MIT License.
 
 using System.Net;
+using System.Text;
 using System.Text.Json.Nodes;
+using Azure.Core;
+using Azure.Mcp.Core.Services.Azure.Tenant;
 using Azure.Mcp.Tools.Monitor.Commands.HealthModels.Entity;
 using Azure.Mcp.Tools.Monitor.Services;
+using Azure.ResourceManager;
 using Microsoft.Mcp.Core.Models;
 using Microsoft.Mcp.Core.Options;
+using Microsoft.Mcp.Core.Services.Azure.Authentication;
 using Microsoft.Mcp.Tests.Client;
 using NSubstitute;
 using NSubstitute.ExceptionExtensions;
@@ -22,6 +27,7 @@ public class EntityGetHealthCommandTests : CommandUnitTestsBase<EntityGetHealthC
     private const string TestResourceGroup = "resourceGroup1";
     private const string TestSubscription = "sub123";
     private const string TestTenant = "tenant123";
+    private const string SupportedApiVersion = "2025-05-01-preview";
 
     [Fact]
     public async Task ExecuteAsync_WithValidParameters_ReturnsEntityHealth()
@@ -287,5 +293,123 @@ public class EntityGetHealthCommandTests : CommandUnitTestsBase<EntityGetHealthC
             Arg.Any<string>(),
             Arg.Is<RetryPolicyOptions>(r => r.DelaySeconds == RetryDelay && r.MaxRetries == MaxRetries),
             Arg.Any<CancellationToken>());
+    }
+
+    [Theory]
+    [InlineData("https://contoso.healthmodels.azure.com")]   // no trailing slash
+    [InlineData("https://contoso.healthmodels.azure.com/")]  // trailing slash
+    public async Task GetEntityHealth_BuildsWellFormedDataplaneUrl(string dataplaneEndpoint)
+    {
+        // Arrange
+        const string entity = "2c139c0b-87d8-4935-ae18-4217431002d2";
+        var handler = new CapturingHttpMessageHandler(request =>
+        {
+            var url = request.RequestUri!.ToString();
+            if (url.Contains("management.azure.com", StringComparison.OrdinalIgnoreCase))
+            {
+                var body = "{\"properties\":{\"dataplaneEndpoint\":\"" + dataplaneEndpoint + "\"}}";
+                return new HttpResponseMessage(HttpStatusCode.OK)
+                {
+                    Content = new StringContent(body, Encoding.UTF8, "application/json")
+                };
+            }
+
+            return new HttpResponseMessage(HttpStatusCode.OK)
+            {
+                Content = new StringContent("""{"history":[]}""", Encoding.UTF8, "application/json")
+            };
+        });
+
+        var service = CreateService(handler);
+
+        // Act
+        var result = await service.GetEntityHealth(
+            entity,
+            "contoso",
+            "rg1",
+            "12345678-1234-1234-1234-123456789012",
+            cancellationToken: TestContext.Current.CancellationToken);
+
+        // Assert
+        Assert.NotNull(result);
+
+        var controlPlaneRequest = Assert.Single(handler.Requests, r => r.RequestUri!.Host == "management.azure.com");
+        Assert.Contains($"api-version={SupportedApiVersion}", controlPlaneRequest.RequestUri!.Query, StringComparison.Ordinal);
+
+        var dataplaneRequest = Assert.Single(handler.Requests, r => r.RequestUri!.Host == "contoso.healthmodels.azure.com");
+        Assert.Equal($"https://contoso.healthmodels.azure.com/api/entities/{entity}/history", dataplaneRequest.RequestUri!.AbsoluteUri);
+    }
+
+    [Fact]
+    public async Task GetEntityHealth_UrlEncodesEntityName()
+    {
+        // Arrange
+        const string entity = "my entity/with special?chars";
+        var handler = new CapturingHttpMessageHandler(request =>
+        {
+            var url = request.RequestUri!.ToString();
+            if (url.Contains("management.azure.com", StringComparison.OrdinalIgnoreCase))
+            {
+                return new HttpResponseMessage(HttpStatusCode.OK)
+                {
+                    Content = new StringContent("""{"properties":{"dataplaneEndpoint":"https://m.healthmodels.azure.com"}}""", Encoding.UTF8, "application/json")
+                };
+            }
+
+            return new HttpResponseMessage(HttpStatusCode.OK)
+            {
+                Content = new StringContent("""{"history":[]}""", Encoding.UTF8, "application/json")
+            };
+        });
+
+        var service = CreateService(handler);
+
+        // Act
+        await service.GetEntityHealth(
+            entity,
+            "m",
+            "rg1",
+            "12345678-1234-1234-1234-123456789012",
+            cancellationToken: TestContext.Current.CancellationToken);
+
+        // Assert
+        var dataplaneRequest = Assert.Single(handler.Requests, r => r.RequestUri!.Host == "m.healthmodels.azure.com");
+        Assert.Equal(
+            $"https://m.healthmodels.azure.com/api/entities/{Uri.EscapeDataString(entity)}/history",
+            dataplaneRequest.RequestUri!.AbsoluteUri);
+    }
+
+    private static MonitorHealthModelService CreateService(CapturingHttpMessageHandler handler)
+    {
+        var tenantService = Substitute.For<ITenantService>();
+
+        var cloudConfig = Substitute.For<IAzureCloudConfiguration>();
+        cloudConfig.ArmEnvironment.Returns(ArmEnvironment.AzurePublicCloud);
+        cloudConfig.CloudType.Returns(AzureCloudConfiguration.AzureCloud.AzurePublicCloud);
+        tenantService.CloudConfiguration.Returns(cloudConfig);
+
+        var credential = Substitute.For<TokenCredential>();
+        credential.GetTokenAsync(Arg.Any<TokenRequestContext>(), Arg.Any<CancellationToken>())
+            .Returns(new ValueTask<AccessToken>(new AccessToken("test-token", DateTimeOffset.UtcNow.AddHours(1))));
+        tenantService.GetTokenCredentialAsync(Arg.Any<string?>(), Arg.Any<CancellationToken>())
+            .Returns(Task.FromResult(credential));
+
+        var httpClientFactory = Substitute.For<IHttpClientFactory>();
+        httpClientFactory.CreateClient(Arg.Any<string>()).Returns(_ => new HttpClient(handler));
+
+        return new MonitorHealthModelService(tenantService, httpClientFactory);
+    }
+
+    private sealed class CapturingHttpMessageHandler(Func<HttpRequestMessage, HttpResponseMessage> responder) : HttpMessageHandler
+    {
+        private readonly Func<HttpRequestMessage, HttpResponseMessage> _responder = responder;
+
+        public List<HttpRequestMessage> Requests { get; } = [];
+
+        protected override Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
+        {
+            Requests.Add(request);
+            return Task.FromResult(_responder(request));
+        }
     }
 }
